requirements.txt
cma==3.2.2
imageio==2.19.3
matplotlib==3.5.2
meshio==5.3.4
networkx==2.8.3
numpy==1.22.4
packaging==21.3
Pillow==9.4.0
pybullet==3.2.5
pybullet_svl==3.1.6.4
pyvista==0.36.1
PyYAML==6.0
requests==2.18.4
scikit_image==0.19.2
scikit_learn==1.2.1
scipy==1.5.2
setuptools
tqdm==4.50.2
transforms3d==0.3.1
trimesh==3.12.5
